platform/functions/text-classification.api/pyproject.toml
[tool.poetry]
name = "text_classification"
version = "0.1.0"
description = ""
authors = ["jonirap <joni.rapoport@gmail.com>"]
readme = "README.md"
include = ["models"]

[tool.poetry.dependencies]
python = "~3.10"
sentence-transformers = "^2.2.2"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl"}
gunicorn = "^20.1.0"
flask = "^2.3.2"
pydantic = "^2.2.1"



[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.3.0"
black = "^23.3.0"
notebook = "^6.5.4"
ipython = "^8.14.0"
tqdm = "^4.65.0"
ipywidgets = "^8.0.6"
autoflake = "^2.2.0"
autopep8 = "^2.0.2"
pep8 = "^1.7.1"
isort = "^5.12.0"



[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
secondary = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
